+++ Sources/ARKit-CoreLocation/Nodes/LocationAnnotationNode.swift
@@ -89,7 +89,8 @@ open class LocationAnnotationNode: LocationNode {
         childNodes.first?.renderingOrder = renderingOrder(fromDistance: distance)
 
         let adjustedDistance = self.adjustedDistance(setup: setup, position: position,
-                                                     locationNodeLocation: nodeLocation, locationManager: locationManager)
+                                                     locationNodeLocation: nodeLocation,
+                                                     locationManager: locationManager)
 
         // The scale of a node with a billboard constraint applied is ignored
         // The annotation subnode itself, as a subnode, has the scale applied to it