+++ changelog.md
@@ -1,5 +1,6 @@
 # Changelog
 - 1.3.1
+    - [PR #235 - Adds support to vertically stack annotations](https://github.com/ProjectDent/ARKit-CoreLocation/pull/235)
 - 1.3.0
    - [PR #255 - Fix minor error with placement of route steps](https://github.com/ProjectDent/ARKit-CoreLocation/pull/255)
    - [PR #249 - Fix Memory Leak](https://github.com/ProjectDent/ARKit-CoreLocation/pull/249)